backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crewai
openai>=1.13.3,<2.0.0
pydantic==2.5.0
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
jinja2==3.1.2
reportlab==4.0.7
supabase==2.0.0
stripe==7.8.0
python-dotenv==1.0.0
asyncpg==0.29.0